Download only if nonexistent.

+++ prepare_offline_install.py
@@ -1,4 +1,5 @@
 import sys
+import os
 import urllib
 
 EZ_SETUP_URL = https://bootstrap.pypa.io/ez_setup.py
@@ -8,10 +9,13 @@ current_path = os.path.dirname(os.path.abspath(__file__))
 
 def run():
     print('Collecting assets...')
+
     # ez_setup.py, for setuptools. This could be useful for later installation
     # of the packages using python setup.py install
     ez_setup_path = os.path.join(current_path,'install_assets','ez_setup.py')
-    urllib.urlretrieve(EZ_SETUP_URL,ez_setup_path)
+    if not os.path.isfile(ez_setup_path):
+        # Download if non existent:
+        urllib.urlretrieve(EZ_SETUP_URL,ez_setup_path)
 
     # This should automatically download sqlite3.dll:
     from idsearch.usqlite3 import sqlite3, is_fts4_supported
